# openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/servers/{serverId}/channels:
    get:
      operationId: ServerChannelController_getChannels
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelWithReadResponseDTO"
        "403":
          description: Forbidden
      security: &a15
        - bearer: []
      summary: Get server channels
      tags: &a16
        - Server Channels
    post:
      operationId: ServerChannelController_createChannel
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChannelRequestDTO"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelResponseDTO"
        "403":
          description: Forbidden
      security: *a15
      summary: Create channel
      tags: *a16
  /api/v1/servers/{serverId}/categories:
    get:
      operationId: ServerChannelController_getCategories
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponseDTO"
        "403":
          description: Forbidden
      security: *a15
      summary: Get server categories
      tags: *a16
    post:
      operationId: ServerChannelController_createCategory
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequestDTO"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponseDTO"
        "403":
          description: Forbidden
      security: *a15
      summary: Create category
      tags: *a16
  /api/v1/servers/{serverId}/channels/reorder:
    patch:
      operationId: ServerChannelController_reorderChannels
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderChannelsRequestDTO"
      responses:
        "200":
          description: Channels reordered
        "403":
          description: Forbidden
      security: *a15
      summary: Reorder channels
      tags: *a16
  /api/v1/servers/{serverId}/channels/{channelId}/stats:
    get:
      operationId: ServerChannelController_getChannelStats
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
        - name: channelId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelStatsResponseDTO"
        "400":
          description: Bad Request
        "403":
          description: Forbidden
        "404":
          description: Channel Not Found
      security: *a15
      summary: Get channel stats
      tags: *a16
  /api/v1/servers/{serverId}/channels/{channelId}:
    patch:
      operationId: ServerChannelController_updateChannel
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
        - name: channelId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChannelRequestDTO"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelResponseDTO"
        "403":
          description: Forbidden
        "404":
          description: Channel Not Found
      security: *a15
      summary: Update channel
      tags: *a16
    delete:
      operationId: ServerChannelController_deleteChannel
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
        - name: channelId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Channel deleted
        "403":
          description: Forbidden
        "404":
          description: Channel Not Found
      security: *a15
      summary: Delete channel
      tags: *a16
  /api/v1/servers/{serverId}/categories/{categoryId}:
    patch:
      operationId: ServerChannelController_updateCategory
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
        - name: categoryId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequestDTO"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponseDTO"
        "403":
          description: Forbidden
        "404":
          description: Category Not Found
      security: *a15
      summary: Update category
      tags: *a16
    delete:
      operationId: ServerChannelController_deleteCategory
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
        - name: categoryId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Category deleted
        "403":
          description: Forbidden
        "404":
          description: Category Not Found
      security: *a15
      summary: Delete category
      tags: *a16
  /api/v1/servers/{serverId}/categories/reorder:
    patch:
      operationId: ServerChannelController_reorderCategories
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderCategoriesRequestDTO"
      responses:
        "200":
          description: Categories reordered
        "403":
          description: Forbidden
      security: *a15
      summary: Reorder categories
      tags: *a16
  /api/v1/servers/{serverId}/channels/{channelId}/permissions:
    get:
      operationId: ServerChannelController_getChannelPermissions
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
        - name: channelId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Permissions retrieved
        "403":
          description: Forbidden
        "404":
          description: Channel Not Found
      security: *a15
      summary: Get channel permissions
      tags: *a16
    patch:
      operationId: ServerChannelController_updateChannelPermissions
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
        - name: channelId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePermissionsRequestDTO"
      responses:
        "200":
          description: Permissions updated
        "403":
          description: Forbidden
        "404":
          description: Channel Not Found
      security: *a15
      summary: Update channel permissions
      tags: *a16
  /api/v1/servers/{serverId}/categories/{categoryId}/permissions:
    get:
      operationId: ServerChannelController_getCategoryPermissions
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
        - name: categoryId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Permissions retrieved
        "403":
          description: Forbidden
        "404":
          description: Category Not Found
      security: *a15
      summary: Get category permissions
      tags: *a16
    patch:
      operationId: ServerChannelController_updateCategoryPermissions
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
        - name: categoryId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePermissionsRequestDTO"
      responses:
        "200":
          description: Permissions updated
        "403":
          description: Forbidden
        "404":
          description: Category Not Found
      security: *a15
      summary: Update category permissions
      tags: *a16
...
info:
  title: Serchat API
  description: The Serchat API description
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    ChannelWithReadResponseDTO:
      type: object
      properties:
        _id:
          type: string
        serverId:
          type: string
        name:
          type: string
        type:
          type: string
          enum: &a33
            - text
            - voice
        description:
          type: string
        position:
          type: number
        categoryId:
          type: string
          nullable: true
        lastMessageAt:
          type: string
          nullable: true
        permissions:
          type: object
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        lastReadAt:
          type: string
          nullable: true
      required:
        - serverId
        - name
        - type
        - position
...
    CreateChannelRequestDTO:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum:
            - text
            - voice
        position:
          type: number
        categoryId:
          type: string
        description:
          type: string
      required:
        - name
    ChannelResponseDTO:
      type: object
      properties:
        _id:
          type: string
        serverId:
          type: string
        name:
          type: string
        type:
          type: string
          enum: *a33
        description:
          type: string
        position:
          type: number
        categoryId:
          type: string
          nullable: true
        lastMessageAt:
          format: date-time
          type: string
        permissions:
          type: object
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - serverId
        - name
        - type
        - position
...
    UpdateCategoryRequestDTO:
      type: object
      properties:
        name:
          type: string
        position:
          type: number
...